refactor(cli): split Codex shim server and handlers
